// packages/shared/src/enums/blood-type.enum.d.ts
export declare enum BloodType {
    A_POSITIVE = "A+",
    A_NEGATIVE = "A-",
    B_POSITIVE = "B+",
    B_NEGATIVE = "B-",
    AB_POSITIVE = "AB+",
    AB_NEGATIVE = "AB-",
    O_POSITIVE = "O+",
    O_NEGATIVE = "O-"
}
export declare enum UserMode {
    DONOR = "donor",
    SEEKER = "seeker",
    BOTH = "both"
}
export declare enum RequestStatus {
    OPEN = "open",
    ACCEPTED = "accepted",
    CANCELLED = "cancelled",
    COMPLETED = "completed"
}
export declare enum Urgency {
    STANDARD = "standard",
    NORMAL = "normal",
    URGENT = "urgent",
    CRITICAL = "critical"
}
export declare enum NotificationType {
    REQUEST_MATCH = "request_match",
    REQUEST_ACCEPTED = "request_accepted",
    ELIGIBILITY_REMINDER = "eligibility_reminder"
}